+++ server/src/graql/gremlin/fragment/LabelFragment.java
@@ -90,8 +90,9 @@ public abstract class LabelFragment extends Fragment {
 
 
     @Override
-    public long estimatedCostAsStartingPoint(TransactionOLTP tx) {
+    public double estimatedCostAsStartingPoint(TransactionOLTP tx) {
         // there's only 1 label in this set, but sum anyway
+        // estimate the total number of things that might be connected by ISA to this label as a heuristic
         long instances = labels().stream()
                 .map(label -> tx.session().keyspaceStatistics().count(tx, label.toString()))
                 .reduce((a,b) -> a+b)
